refactor(architecture): add interfaces dir && merge code
