+++ src/renderer/interfaces/containers/iThumbnailRequest.ts
