+++ src/renderer/interfaces/services/iMediaStorable.ts
@@ -1,0 +1,4 @@
+export interface IMediaStorable {
+  generatePathBy(mediaHash: string, tag: string): Promise<string | null>
+  getImageBy(mediaHash: string, tag: string): Promise<string | null>
+}
