--- src/renderer/services/media/index.ts
@@ -1,4 +1,0 @@
-export interface IMediaStorable {
-  generatePathBy(mediaHash: string, tag: string): Promise<string | null>
-  getImageBy(mediaHash: string, tag: string): Promise<string | null>
-}
